// DXC_Frontend/src/api/models/feedbackResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { FeedbackResponseAttachmentDto } from './feedbackResponseAttachmentDto'

export interface FeedbackResponseDto {
  id?: number
  publicId?: string
  departmentId?: number
  /** @nullable */
  departmentCode: string | null
  /** @nullable */
  departmentName: string | null
  /** @nullable */
  responseContent: string | null
  /** @nullable */
  responseAttachments?: string | null
  /** @nullable */
  isApproved?: boolean | null
  /** @nullable */
  approvedByUserPublicId?: string | null
  /** @nullable */
  approvedAt?: string | null
  /** @nullable */
  approvalNote?: string | null
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  attachments?: FeedbackResponseAttachmentDto[] | null
}